+++ src/Stateless/StateMachine.cs
@@ -10,7 +10,7 @@ namespace Stateless
     /// </summary>
     public enum FiringMode
     {
-        /// <summary> Use immediate mode when the queing of trigger events are not needed. Care must be taken when using this mode, as there is no run-to-completion guaranteed.</summary>
+        /// <summary> Use immediate mode when the queuing of trigger events are not needed. Care must be taken when using this mode, as there is no run-to-completion guaranteed.</summary>
         Immediate,
         /// <summary> Use the queued Fire-ing mode when run-to-completion is required. This is the recommended mode.</summary>
         Queued
@@ -28,7 +28,8 @@ namespace Stateless
         private readonly Func<TState> _stateAccessor;
         private readonly Action<TState> _stateMutator;
         private UnhandledTriggerAction _unhandledTriggerAction;
-        private OnTransitionedEvent _onTransitionedEvent;
+        private readonly OnTransitionedEvent _onTransitionedEvent;
+        private readonly OnTransitionedEvent _onTransitionCompletedEvent;
         private readonly FiringMode _firingMode;
 
         private class QueuedTrigger
@@ -87,12 +88,13 @@ namespace Stateless
 
 
         /// <summary>
-        /// Default constuctor
+        /// Default constructor
         /// </summary>
         StateMachine()
         {
             _unhandledTriggerAction = new UnhandledTriggerAction.Sync(DefaultUnhandledTriggerAction);
             _onTransitionedEvent = new OnTransitionedEvent();
+            _onTransitionCompletedEvent = new OnTransitionedEvent();
         }
 
         /// <summary>
@@ -216,6 +218,36 @@ namespace Stateless
         /// Actions associated with leaving the current state and entering the new one
         /// will be invoked.
         /// </summary>
+        /// <param name="trigger">The trigger to fire.</param>
+        /// <param name="args">A variable-length parameters list containing arguments. </param>
+        /// <exception cref="System.InvalidOperationException">The current state does
+        /// not allow the trigger to be fired.</exception>
+        public void Fire(TriggerWithParameters trigger, params object[] args)
+        {
+            if (trigger == null) throw new ArgumentNullException(nameof(trigger));
+            InternalFire(trigger.Trigger, args);
+        }
+
+        /// <summary>
+        /// Specify the arguments that must be supplied when a specific trigger is fired.
+        /// </summary>
+        /// <param name="trigger">The underlying trigger value.</param>
+        /// <param name="argumentTypes">The argument types expected by the trigger.</param>
+        /// <returns>An object that can be passed to the Fire() method in order to
+        /// fire the parameterised trigger.</returns>
+        public TriggerWithParameters SetTriggerParameters(TTrigger trigger, params Type[] argumentTypes)
+        {
+            var configuration = new TriggerWithParameters(trigger, argumentTypes);
+            SaveTriggerConfiguration(configuration);
+            return configuration;
+        }
+
+        /// <summary>
+        /// Transition from the current state via the specified trigger.
+        /// The target state is determined by the configuration of the current state.
+        /// Actions associated with leaving the current state and entering the new one
+        /// will be invoked.
+        /// </summary>
         /// <typeparam name="TArg0">Type of the first trigger argument.</typeparam>
         /// <param name="trigger">The trigger to fire.</param>
         /// <param name="arg0">The first argument.</param>
@@ -268,7 +300,7 @@ namespace Stateless
         }
 
         /// <summary>
-        /// Activates current state. Actions associated with activating the currrent state
+        /// Activates current state. Actions associated with activating the current state
         /// will be invoked. The activation is idempotent and subsequent activation of the same current state
         /// will not lead to re-execution of activation callbacks.
         /// </summary>
@@ -279,7 +311,7 @@ namespace Stateless
         }
 
         /// <summary>
-        /// Deactivates current state. Actions associated with deactivating the currrent state
+        /// Deactivates current state. Actions associated with deactivating the current state
         /// will be invoked. The deactivation is idempotent and subsequent deactivation of the same current state
         /// will not lead to re-execution of deactivation callbacks.
         /// </summary>
@@ -415,11 +447,14 @@ namespace Stateless
 
                 _onTransitionedEvent.Invoke(transition);
                 representation = EnterState(newRepresentation, transition, args);
+                _onTransitionCompletedEvent.Invoke(transition);
+
             }
             else
             {
                 _onTransitionedEvent.Invoke(transition);
                 representation = EnterState(newRepresentation, transition, args);
+                _onTransitionCompletedEvent.Invoke(transition);
             }
             State = representation.UnderlyingState;
         }
@@ -436,10 +471,13 @@ namespace Stateless
             var representation = EnterState(newRepresentation, transition, args);
 
             // Check if state has changed by entering new state (by fireing triggers in OnEntry or such)
-            if (representation.UnderlyingState.Equals(State)) return;
+            if (!representation.UnderlyingState.Equals(State))
+            {
+                // The state has been changed after entering the state, must update current state to new one
+                State = representation.UnderlyingState;
+            }
 
-            // The state has been changed after entering the state, must update current state to new one
-            State = representation.UnderlyingState;
+            _onTransitionCompletedEvent.Invoke(new Transition(transition.Source, State, transition.Trigger, transition.Parameters));
         }
 
         private StateRepresentation EnterState(StateRepresentation representation, Transition transition, object [] args)
@@ -467,6 +505,9 @@ namespace Stateless
 
                 var initialTransition = new InitialTransition(transition.Source, representation.InitialTransitionTarget, transition.Trigger, args);
                 representation = GetRepresentation(representation.InitialTransitionTarget);
+
+                // Alert all listeners of initial state transition
+                _onTransitionedEvent.Invoke(new Transition(transition.Destination, initialTransition.Destination, transition.Trigger, transition.Parameters));
                 representation = EnterState(representation, initialTransition, args);
             }
 
@@ -598,7 +639,7 @@ namespace Stateless
         }
 
         /// <summary>
-        /// Registers a callback that will be invoked every time the statemachine
+        /// Registers a callback that will be invoked every time the state machine
         /// transitions from one state into another.
         /// </summary>
         /// <param name="onTransitionAction">The action to execute, accepting the details
@@ -607,6 +648,19 @@ namespace Stateless
         {
             if (onTransitionAction == null) throw new ArgumentNullException(nameof(onTransitionAction));
             _onTransitionedEvent.Register(onTransitionAction);
+        }
+
+        /// <summary>
+        /// Registers a callback that will be invoked every time the statemachine
+        /// transitions from one state into another and all the OnEntryFrom etc methods
+        /// have been invoked
+        /// </summary>
+        /// <param name="onTransitionAction">The action to execute, accepting the details
+        /// of the transition.</param>
+        public void OnTransitionCompleted(Action<Transition> onTransitionAction)
+        {
+            if (onTransitionAction == null) throw new ArgumentNullException(nameof(onTransitionAction));
+            _onTransitionCompletedEvent.Register(onTransitionAction);
         }
     }
 }