+++ test/Stateless.Tests/DynamicTriggerBehaviourFixture.cs
@@ -1,4 +1,5 @@
-﻿using Xunit;
+﻿using System.Linq;
+using Xunit;
 
 namespace Stateless.Tests
 {
@@ -27,6 +28,20 @@ namespace Stateless.Tests
             sm.Fire(trigger, 1);
 
             Assert.Equal(State.B, sm.State);
+        }
+
+        [Fact]
+        public void Sdfsf()
+        {
+            var sm = new StateMachine<State, Trigger>(State.A);
+            var trigger = sm.SetTriggerParameters<int>(Trigger.X);
+            sm.Configure(State.A)
+                .PermitDynamicIf(trigger, (i) => i == 1 ? State.C : State.B, (i) => i == 1 ? true : false);
+
+            // Should not throw
+            sm.GetPermittedTriggers().ToList();
+
+            sm.Fire(trigger, 1);
         }
     }
 }