+++ test/Stateless.Tests/StateMachineFixture.cs
@@ -336,7 +336,7 @@ namespace Stateless.Tests
         }
 
         [Fact]
-        public void WhenATransitionOccurs_TheOnTransitionEventFires()
+        public void WhenATransitionOccurs_TheOnTransitionedEventFires()
         {
             var sm = new StateMachine<State, Trigger>(State.B);
 
@@ -356,10 +356,33 @@ namespace Stateless.Tests
         }
 
         [Fact]
-        public void TheOnTransitionEventFiresBeforeTheOnEntryEvent()
+        public void WhenATransitionOccurs_TheOnTransitionCompletedEventFires()
         {
             var sm = new StateMachine<State, Trigger>(State.B);
-            var expectedOrdering = new List<string> { "OnExit", "OnTransitioned", "OnEntry" };
+
+            sm.Configure(State.B)
+                .Permit(Trigger.X, State.A);
+
+            StateMachine<State, Trigger>.Transition transition = null;
+            sm.OnTransitionCompleted(t => transition = t);
+
+            sm.Fire(Trigger.X);
+
+            Assert.NotNull(transition);
+            Assert.Equal(Trigger.X, transition.Trigger);
+            Assert.Equal(State.B, transition.Source);
+            Assert.Equal(State.A, transition.Destination);
+            Assert.Equal(new object[0], transition.Parameters);
+        }
+
+        /// <summary>
+        /// The expected ordering is OnExit, OnTransitioned, OnEntry, OnTransitionCompleted
+        /// </summary>
+        [Fact]
+        public void TheOnTransitionedEventFiresBeforeTheOnEntryEventAndOnTransitionCompletedFiresAfterwards()
+        {
+            var sm = new StateMachine<State, Trigger>(State.B);
+            var expectedOrdering = new List<string> { "OnExit", "OnTransitioned", "OnEntry", "OnTransitionCompleted" };
             var actualOrdering = new List<string>();
 
             sm.Configure(State.B)
@@ -370,6 +393,7 @@ namespace Stateless.Tests
                 .OnEntry(() => actualOrdering.Add("OnEntry"));
 
             sm.OnTransitioned(t => actualOrdering.Add("OnTransitioned"));
+            sm.OnTransitionCompleted(t => actualOrdering.Add("OnTransitionCompleted"));
 
             sm.Fire(Trigger.X);
 
@@ -381,7 +405,7 @@ namespace Stateless.Tests
         }
 
         [Fact]
-        public void WhenATransitionOccurs_WithAParameterizedTrigger_TheOnTransitionEventFires()
+        public void WhenATransitionOccurs_WithAParameterizedTrigger_TheOnTransitionedEventFires()
         {
             var sm = new StateMachine<State, Trigger>(State.B);
             var triggerX = sm.SetTriggerParameters<string>(Trigger.X);
@@ -404,7 +428,30 @@ namespace Stateless.Tests
         }
 
         [Fact]
-        public void WhenATransitionOccurs_WithAParameterizedTrigger_WithMultipleParameters_TheOnTransitionEventFires()
+        public void WhenATransitionOccurs_WithAParameterizedTrigger_TheOnTransitionCompletedEventFires()
+        {
+            var sm = new StateMachine<State, Trigger>(State.B);
+            var triggerX = sm.SetTriggerParameters<string>(Trigger.X);
+
+            sm.Configure(State.B)
+                .Permit(Trigger.X, State.A);
+
+            StateMachine<State, Trigger>.Transition transition = null;
+            sm.OnTransitionCompleted(t => transition = t);
+
+            string parameter = "the parameter";
+            sm.Fire(triggerX, parameter);
+
+            Assert.NotNull(transition);
+            Assert.Equal(Trigger.X, transition.Trigger);
+            Assert.Equal(State.B, transition.Source);
+            Assert.Equal(State.A, transition.Destination);
+            Assert.Equal(1, transition.Parameters.Count());
+            Assert.Equal(parameter, transition.Parameters[0]);
+        }
+
+        [Fact]
+        public void WhenATransitionOccurs_WithAParameterizedTrigger_WithMultipleParameters_TheOnTransitionedEventFires()
         {
             var sm = new StateMachine<State, Trigger>(State.B);
             var triggerX = sm.SetTriggerParameters<string, int, bool>(Trigger.X);
@@ -414,6 +461,33 @@ namespace Stateless.Tests
 
             StateMachine<State, Trigger>.Transition transition = null;
             sm.OnTransitioned(t => transition = t);
+
+            string firstParameter = "the parameter";
+            int secondParameter = 99;
+            bool thirdParameter = true;
+            sm.Fire(triggerX, firstParameter, secondParameter, thirdParameter);
+
+            Assert.NotNull(transition);
+            Assert.Equal(Trigger.X, transition.Trigger);
+            Assert.Equal(State.B, transition.Source);
+            Assert.Equal(State.A, transition.Destination);
+            Assert.Equal(3, transition.Parameters.Count());
+            Assert.Equal(firstParameter, transition.Parameters[0]);
+            Assert.Equal(secondParameter, transition.Parameters[1]);
+            Assert.Equal(thirdParameter, transition.Parameters[2]);
+        }
+
+        [Fact]
+        public void WhenATransitionOccurs_WithAParameterizedTrigger_WithMultipleParameters_TheOnTransitionCompletedEventFires()
+        {
+            var sm = new StateMachine<State, Trigger>(State.B);
+            var triggerX = sm.SetTriggerParameters<string, int, bool>(Trigger.X);
+
+            sm.Configure(State.B)
+                .Permit(Trigger.X, State.A);
+
+            StateMachine<State, Trigger>.Transition transition = null;
+            sm.OnTransitionCompleted(t => transition = t);
 
             string firstParameter = "the parameter";
             int secondParameter = 99;