+++ CHANGELOG.md
@@ -4,6 +4,56 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 5.2.0
+### Added
+ - Added support for net50
+### Changed
+ - Only run tests under net50
+ - Non-code change: Switch to using Github Actions
+### Fixed
+ - Fixed the bug "IgnoreIf Guard description is raised when PermitReentryIf Guard Fails". [#422]
+
+## 5.1.8
+### Fixed
+- Fixed bug "When adding .PermitDynamicIf, call .PermitTriggers throw NullReferenceException" [#416]
+
+## 5.1.7 - 2021.01.10
+### Fixed
+ - Fixed nullReferenceException when getting permitted triggers on a dynamic transition. [#413]
+
+## 5.1.7 - 2021.01.10
+### Fixed
+ - Fixed OnTransitionCompletedEvent, it now has the correct Destination state when there is an initial transition. [#413]
+
+## 5.1.6 - 2020.11.21
+### Fixed
+ - Added support for new OnTransitionCompletedEvent, which is run after all OnExit / OnEntry methods have run [#394]
+
+## 5.1.5 - 2020.11.16
+### Changed
+- Fixed spelling errors
+- Fixed analyser warnings (mostly null checks )
+### Fixed
+ - Added CLSCompliant attribute, which went missing in November 2019 [#401]
+
+## 5.1.5 - 2020.11.16
+### Changed
+- Fixed spelling errors
+- Fixed analyser warnings (mostly null checks )
+### Fixed
+ - Added CLSCompliant attribute, which went missing in November 2019 [#401]
+
+## 5.1.4 - 2020.11.07
+### Fixed
+ - Fixed Unmet guard collection not set by OnUnhandledTrigger if transition guard of substate fails [#398] (Thanks to the awesome DeepakParamkusam)
+ 
+## 5.1.3 - 2020.08.12
+### Fixed
+ - Fixed ambiguos guard function call when trigger is a state [#380]
+ 
+## 5.1.2 - 2020.04.27
+### Fixed
+ - Fixed wrong onTransitionEvent ordering in Async firing [#372]
 
 ## [5.1.1] - 2020.04.03
 ### Changed
@@ -111,6 +161,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 ### Fixed
 
+[#422]: https://github.com/dotnet-state-machine/stateless/issues/422
+[#416]: https://github.com/dotnet-state-machine/stateless/issues/416
+[#413]: https://github.com/dotnet-state-machine/stateless/issues/413
+[#394]: https://github.com/dotnet-state-machine/stateless/issues/394
+[#401]: https://github.com/dotnet-state-machine/stateless/issues/401
+[#398]: https://github.com/dotnet-state-machine/stateless/issues/398
+[#380]: https://github.com/dotnet-state-machine/stateless/issues/380
+[#373]: https://github.com/dotnet-state-machine/stateless/issues/372
 [#365]: https://github.com/dotnet-state-machine/stateless/pull/365
 [#272]: https://github.com/dotnet-state-machine/stateless/issues/272
 [#275]: https://github.com/dotnet-state-machine/stateless/issues/275